+++ CHANGELOG.md
@@ -48,6 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed calls to removed Ansible playbooks ``enable-dns-conflict-hosts.yml``
   and ``disable-dns-conflict-hosts.yml`` during ``ncn-power`` stage of
   ``sat bootsys``.
+- Removed the ``bgp-check`` stage of ``sat bootsys {boot,shutdown}`` due to
+  removal of underlying Ansible playbook that implemented this stage.
 
 ### Fixed
 - Removed error message and 'ERROR' value for Slurm version in system